lint and more refactor on utils
