+++ .jshintrc
@@ -1,5 +1,5 @@
 {
-    "esversion": 6,
+    "esversion": 9,
     "asi" : true,
     "loopfunc": true,
     "globals": {
