+++ lib/classes/application.js
@@ -13,9 +13,9 @@ var async = require("async");
 var _ = require("underscore"); // General utility functions external module
 
 // Internals dependencies
-var pub = require("./../easyrtc_public_obj"); // EasyRTC public object
+var pub = require("./../index"); // EasyRTC public object
 // TODO pub.socketServer
-var easyrtc_utils = require("./../easyrtc_utils"); // General utility functions local module
+var easyrtc_utils = require("./../utils"); // General utility functions local module
 
 // Internals Models
 const Connection = require('./connection');
