+++ lib/classes/room.js
@@ -9,7 +9,7 @@
  */
 
 // Internals dependencies
-const easyrtc_utils = require("./../easyrtc_utils"); // General utility functions local module
+const easyrtc_utils = require("./../utils"); // General utility functions local module
 
 // TODO migreate to ES6 or TypeScript
 module.exports = function Room(appName, roomName, appObj, server) {
