+++ lib/classes/server.js
@@ -13,7 +13,7 @@ const async = require("async");
 const _ = require("underscore"); // General utility functions external module
 
 // Internals dependencies
-const easyrtc_utils = require("./../easyrtc_utils"); // General utility functions local module
+const easyrtc_utils = require("./../utils"); // General utility functions local module
 
 const EventEmitter = require('./event-emitter');
 const Application = require('./application');
@@ -22,23 +22,27 @@ const defaultOptions = require("./server/default-options"); // EasyRTC global va
 // TODO migreate to ES6 or TypeScript
 const Server = class Server {
 
-    // TODO readonly
-    serverStartOn = Date.now();
-
-    // TODO readonly
-    option = easyrtc_utils.deepCopy(defaultOptions);
-
-    // TODO map readonly
-    app = {};
-
-    socketServer = null;
-
     /**
      * EasyRTC Event handling object which contain most methods for interacting with EasyRTC events.
      * For convenience, this class has also been attached to the application, connection, session, and room classes.
      * @class
      */
-    events = new EventEmitter();
+
+    constructor(socketServer) {
+
+        // TODO readonly
+        this.serverStartOn = Date.now();
+
+        // TODO readonly
+        this.option = easyrtc_utils.deepCopy(defaultOptions);
+
+        // TODO map readonly
+        this.app = {};
+
+        // Running the default listeners to initialize the events
+        this.events = new EventEmitter();
+        this.events.setDefaultListeners();
+    }
 
     get listeners() {
         return this.events.listeners;
@@ -72,11 +76,6 @@ const Server = class Server {
             this.events.removeAllListeners(event);
             this.events.setDefaultListener(event);
         }
-    }
-
-    constructor(socketServer) {
-        // Running the default listeners to initialize the events
-        this.events.setDefaultListeners();
     }
 
     /**
@@ -179,7 +178,7 @@ const Server = class Server {
      *
      * @param       {function(?Error, Number)} callback Callback with error and array containing all easyrtcids.
      */
-    getConnectionCount = function(callback) {
+    getConnectionCount(callback) {
         callback(null, getConnectionCountSync());
     }
 
@@ -196,7 +195,7 @@ const Server = class Server {
             }
         }
         return connectionCount;
-    };
+    }
 
     /**
      * Gets app object for application which has an authenticated client with a given easyrtcid
@@ -204,7 +203,7 @@ const Server = class Server {
      * @param       {String} easyrtcid      Unique identifier for an EasyRTC connection.
      * @param       {function(?Error, Object=)} callback Callback with error and application object
      */
-    getAppWithEasyrtcid = function(easyrtcid, callback) {
+    getAppWithEasyrtcid(easyrtcid, callback) {
         for (var appName in this.app) {
             if (this.app.hasOwnProperty(appName)) {
                 if (
@@ -218,7 +217,7 @@ const Server = class Server {
         }
         easyrtc_utils.logWarning("Can not find connection [" + easyrtcid + "]");
         callback(new easyrtc_utils.ConnectionWarning("Can not find connection [" + easyrtcid + "]"));
-    };
+    }
 
     /**
      * Gets connection object for connection which has an authenticated client with a given easyrtcid
@@ -226,7 +225,7 @@ const Server = class Server {
      * @param       {string} easyrtcid      EasyRTC unique identifier for a socket connection.
      * @param       {function(?Error, Object=)} callback Callback with error and connection object
      */
-    getConnectionWithEasyrtcid = function(easyrtcid, callback) {
+    getConnectionWithEasyrtcid(easyrtcid, callback) {
         for (var appName in this.app) {
             if (this.app.hasOwnProperty(appName)) {
                 if (
