+++ lib/classes/server/default-listeners.js
@@ -14,7 +14,7 @@ const async = require("async"); // Asynchronous calls external module
 const _ = require("underscore"); // General utility functions external module
 
 // Internals dependencies
-const pub = require("./../../easyrtc_public_obj"); // EasyRTC public object
+const pub = require("./../../index"); // EasyRTC public object
 
 /**
  * Event listeners used by EasyRTC. Many of these can be overridden using server options. The interfaces should be used as a guide for creating new listeners.
@@ -1115,8 +1115,7 @@ eventListener.onEmitReturnToken = function(connectionObj, socketCallback, next) 
 
                 // TODO config name for easyrtcsid ?
                 tokenMsg.msgData.sessionData = {
-                    "easyrtcsid":
-                    sessionObj.getEasyrtcsid()
+                    "easyrtcsid": sessionObj.getEasyrtcsid()
                 };
 
                 // Get session fields
@@ -1609,22 +1608,6 @@ eventListener.onStartup = function(next) {
         next = pub.util.nextToNowhere;
     }
 
-    const easyrtcRoot = __dirname + "/../../../";
-
-    // TODO use this instead of duplcate code
-    var sendResponse = function(req, res, filePath) {
-        pub.util.sendSessionCookie(req, res);
-        (res.sendFile||res.sendfile).call(res, filePath,  {
-            root: easyrtcRoot
-        }, function (err) {
-            console.log(err);
-            if (err && err.status && res && !res._headerSent) {
-                res.status(404);
-                res.end();
-            }
-        });
-    }
-
     pub.util.logDebug("Running func 'onStartup'");
     async.waterfall([
         function(callback) {
@@ -1636,7 +1619,8 @@ eventListener.onStartup = function(next) {
 
                 pub.httpApp.get(pub.getOption("demosPublicFolder") + "/*", function(req, res) {
                     const filePath = "./demos/" + (req.params[0] ? req.params[0] : "index.html");
-                    sendResponse(req, res, filePath);
+                    pub.util.sendSessionCookie(req, res);
+                    pub.util.sendFileResponse(req, res, filePath);
                 });
                 // Forward people who forget the trailing slash to the folder.
                 pub.httpApp.get(pub.getOption("demosPublicFolder"), function(req, res) {res.redirect(pub.getOption("demosPublicFolder") + "/");});
@@ -1646,18 +1630,18 @@ eventListener.onStartup = function(next) {
             if (pub.getOption("apiEnable")) {
                 pub.util.logDebug("Setting up API files to be accessed from '" + pub.getOption("apiPublicFolder") + "/'");
 
-                pub.httpApp.get(pub.getOption("apiPublicFolder") + "/easyrtc.js",     function(req, res) { sendResponse(req, res, "api/easyrtc.js"); });
-                pub.httpApp.get(pub.getOption("apiPublicFolder") + "/easyrtc_ft.js",  function(req, res) { sendResponse(req, res, "api/easyrtc_ft.js"); });
-                pub.httpApp.get(pub.getOption("apiPublicFolder") + "/easyrtc.css",    function(req, res) { sendResponse(req, res, "api/easyrtc.css"); });
+                pub.httpApp.get(pub.getOption("apiPublicFolder") + "/easyrtc.js",     function(req, res) { pub.util.sendFileResponse(req, res, "api/easyrtc.js"); });
+                pub.httpApp.get(pub.getOption("apiPublicFolder") + "/easyrtc_ft.js",  function(req, res) { pub.util.sendFileResponse(req, res, "api/easyrtc_ft.js"); });
+                pub.httpApp.get(pub.getOption("apiPublicFolder") + "/easyrtc.css",    function(req, res) { pub.util.sendFileResponse(req, res, "api/easyrtc.css"); });
                 pub.httpApp.get(pub.getOption("apiPublicFolder") + "/img/*", function(req, res) {
                     const filePath = "./api/img/" + (req.params[0] ? req.params[0] : "index.html");
-                    sendResponse(req, res, filePath);
+                    pub.util.sendFileResponse(req, res, filePath);
                 });
 
                 if (pub.getOption("apiLabsEnable")) {
                     pub.httpApp.get(pub.getOption("apiPublicFolder") + "/labs/*", function(req, res) {
                         const filePath = "./api/labs/" + (req.params[0] ? req.params[0] : "index.html");
-                        sendResponse(req, res, filePath);
+                        pub.util.sendFileResponse(req, res, filePath);
                     });
                 }
             }
@@ -1665,8 +1649,8 @@ eventListener.onStartup = function(next) {
             // Transition - Old locations of EasyRTC API files
             if (pub.getOption("apiEnable") && pub.getOption("apiOldLocationEnable")) {
                 pub.util.logWarning("Enabling listening for API files in older depreciated location.");
-                pub.httpApp.get("/js/easyrtc.js", function(req, res) { sendResponse(req, res, "api/easyrtc.js"); });
-                pub.httpApp.get("/css/easyrtc.css", function(req, res) { sendResponse(req, res, "api/easyrtc.css"); });
+                pub.httpApp.get("/js/easyrtc.js", function(req, res) { pub.util.sendFileResponse(req, res, "api/easyrtc.js"); });
+                pub.httpApp.get("/css/easyrtc.css", function(req, res) { pub.util.sendFileResponse(req, res, "api/easyrtc.css"); });
             }
 
             callback(null);
