+++ lib/utils/easyrtc.js
@@ -10,46 +10,20 @@
  */
 
 // Externals dependencies
-var async = require("async");
-var _ = require("underscore"); // General utility functions external module
+const util = require("util");
+const async = require("async");
+const _ = require("underscore"); // General utility functions external module
 
 // Internals dependencies
-var utils = require("./utils");
-var utils_error = require("./utils/error");
-var utils_log = require("./utils/log");
+const utils_error = require("./error");
+const utils_log = require("./log");
 
 /**
  *  Object to hold EasyRTC Utility methods and classes.
  *
  * @class
  */
-var easyrtc_utils = module.exports = {
-    ...utils_error,
-    ...utils_log,
-
-    get logLevel() {
-        return utils_log.logLevel;
-    },
-    set logLevel(level) {
-        return utils_log.logLevel = level;
-    },
-    get logCallback() {
-        return utils_log.logCallback;
-    },
-    set logCallback(callback) {
-        utils_log.logCallback = callback;
-    }
-};
-
-/**
- * Performs a deep copy of an object, returning the duplicate.
- * Do not use on objects with circular references.
- *
- * @function
- * @param       {Object} input          Input variable (or object) to be copied.
- * @returns     {Object}                New copy of variable.
- */
-easyrtc_utils.deepCopy = utils.deepCopy;
+const easyrtc_utils = module.exports;
 
 /*
  * Return a random string of characters
@@ -217,7 +191,7 @@ easyrtc_utils.sendSessionCookie = function(req, res) {
     }
 
     // TODO allow easyrtcsid value for cookie name options
-    if (req.sessionID && (!req.cookies || !req.cookies.easyrtcsid || req.cookieseasyrtcsid !== req.sessionID)) {
+    if (req.sessionID && (!req.cookies || !req.cookies.easyrtcsid || req.cookies.easyrtcsid !== req.sessionID)) {
         try {
             easyrtc_utils.logDebug("Sending easyrtcsid cookie [" + req.sessionID + "] to [" + req.ip + "] for request [" + req.url + "]");
             res.cookie("easyrtcsid", req.sessionID, {
@@ -229,6 +203,28 @@ easyrtc_utils.sendSessionCookie = function(req, res) {
         }
     }
 };
+
+/**
+ * Will attempt to deliver an EasyRTC file via a sendFile.
+ *
+ * @param       {Object} req            Http request object
+ * @param       {Object} res            Http result object
+ * @param       {Object} filePath       Path of file to send
+ */
+easyrtc_utils.sendFileResponse = function(req, res, filePath) {
+
+    const easyrtcRoot = __dirname + "/../";
+
+    (res.sendFile||res.sendfile).call(res, filePath,  {
+        root: easyrtcRoot
+    }, function (err) {
+        if (err && err.status && res && !res._headerSent) {
+            pub.util.logError("Send file error", err);
+            res.status(404);
+            res.end();
+        }
+    });
+}
 
 easyrtc_utils.isValidIncomingType = (type, msg, callback) => {
     // All messages follow the basic structure
@@ -615,23 +611,68 @@ easyrtc_utils.isValidIncomingMessage = function(type, msg, obj, callback) {
     switch (type) {
         case "easyrtcAuth":
             easyrtc_utils.isValidIncomingAuth(type, msg, obj, callback);
-            return; /* jshint ignore:line */
-
-        break; /* jshint ignore:line */
+            return;
 
         case "easyrtcCmd":
             easyrtc_utils.isValidIncomingCmd(type, msg, obj, callback);
-            return; /* jshint ignore:line */
-
-        break;
+            return;
 
         case "easyrtcMsg":
             easyrtc_utils.isValidIncomingMsg(type, msg, obj, callback);
-            return; /* jshint ignore:line */
-        break;
+            return;
 
         default:
             callback(null, false, "MSG_REJECT_BAD_TYPE");
             return;
     }
+};
+
+
+/**
+ * Performs a deep copy of an object, returning the duplicate.
+ * Do not use on objects with circular references.
+ *
+ * @param       {Object} input          Input variable (or object) to be copied.
+ * @returns     {Object}                New copy of variable.
+ */
+easyrtc_utils.deepCopy = function(input) {
+
+    if (
+        input === null || input === undefined ||
+            typeof input !== "object" ||
+                (input.constructor !== Object && input.constructor !== Array)
+    ) {
+        return input;
+    }
+
+    if (
+        input.constructor === Boolean ||
+            input.constructor === Date ||
+                input.constructor === Function ||
+                    input.constructor === Number ||
+                        input.constructor === RegExp ||
+                            input.constructor === String
+    ) {
+        return new input.constructor(input);
+    }
+
+    var copy;
+    if (input instanceof Array) {
+        copy = [];
+        for (var i = 0, len = input.length; i < len; i++) {
+            copy[i] = easyrtc_utils.deepCopy(input[i]);
+        }
+        return copy;
+    }
+
+    if (input instanceof Object) {
+        copy = {};
+        for (var key in input) {
+            if (input.hasOwnProperty(key)) {
+                copy[key] = easyrtc_utils.deepCopy(input[key]);
+            }
+        }
+        return copy;
+    }
+    return null;
 };
