+++ lib/utils/log.js
@@ -8,7 +8,7 @@
  * @license     BSD v2, see LICENSE file in module root folder.
  */
 
-var easyrtc_log = module.exports = {
+const easyrtc_log = module.exports = {
 
     // The minimum log level to show. (debug|info|warning|error|none)
     logLevel: 'debug',
